Fixed "Proxied: 0" bug on slow hardware
