+++ src/background/handlers/onProxyRequest.ts
@@ -1,4 +1,4 @@
-import { Proxy } from "webextension-polyfill";
+import browser, { Proxy } from "webextension-polyfill";
 import findChannelFromTwitchTvUrl from "../../common/ts/findChannelFromTwitchTvUrl";
 import findChannelFromUsherUrl from "../../common/ts/findChannelFromUsherUrl";
 import findChannelFromVideoWeaverUrl from "../../common/ts/findChannelFromVideoWeaverUrl";
@@ -109,9 +109,12 @@ export default async function onProxyRequest(
 
   // Video Weaver requests.
   if (proxyVideoWeaverRequest && videoWeaverHostRegex.test(host)) {
+    const tab =
+      details.tabId !== -1 ? await browser.tabs.get(details.tabId) : null;
+    const tabUrl = tab?.url;
     const channelName =
       findChannelFromVideoWeaverUrl(details.url) ??
-      findChannelFromTwitchTvUrl(details.documentUrl);
+      findChannelFromTwitchTvUrl(tabUrl);
     if (isChannelWhitelisted(channelName)) {
       console.log(`✋ Channel '${channelName}' is whitelisted.`);
       return { type: "direct" };
